Add support for schroot. Closes #530733

+++ piuparts.py
@@ -146,6 +146,7 @@ class Settings:
         self.basetgz = None
         self.savetgz = None
         self.lvm_volume = None
+        self.schroot = None
         self.end_meta = None
         self.save_end_meta = None
         self.skip_minimize = True
@@ -722,7 +723,8 @@ class Chroot:
 
     def create(self, temp_tgz = None):
         """Create a chroot according to user's wishes."""
-        self.create_temp_dir()
+        if not settings.schroot:
+            self.create_temp_dir()
         cid = do_on_panic(self.remove)
 
         if temp_tgz:
@@ -731,13 +733,16 @@ class Chroot:
             self.unpack_from_tgz(settings.basetgz)
         elif settings.lvm_volume:
             self.setup_from_lvm(settings.lvm_volume)
+        elif settings.schroot:
+            self.setup_from_schroot(settings.schroot)
         else:
             self.setup_minimal_chroot()
 
-        self.mount_proc()
-        self.mount_selinux()
+        if not settings.schroot:
+            self.mount_proc()
+            self.mount_selinux()
         self.configure_chroot()
-        if settings.basetgz:
+        if settings.basetgz or settings.schroot:
             self.run(["apt-get", "-yf", "dist-upgrade"])
         self.minimize()
 
@@ -764,16 +769,24 @@ class Chroot:
     def remove(self):
         """Remove a chroot and all its contents."""
         if not settings.keep_tmpdir and os.path.exists(self.name):
-            self.unmount_proc()
-            self.unmount_selinux()
+            if not settings.schroot:
+                self.unmount_proc()
+                self.unmount_selinux()
             if settings.lvm_volume:
                 logging.debug('Unmounting and removing LVM snapshot %s' % self.lvm_snapshot_name)
                 run(['umount', self.name])
                 run(['lvremove', '-f', self.lvm_snapshot])
-            shutil.rmtree(self.name)
-            logging.debug("Removed directory tree at %s" % self.name)
+            if settings.schroot:
+                logging.debug("Terminate schroot session '%s'" % self.name)
+                run(['schroot', '--end-session', '--chroot', self.schroot_session])
+            if not settings.schroot:
+                shutil.rmtree(self.name)
+                logging.debug("Removed directory tree at %s" % self.name)
         elif settings.keep_tmpdir:
-            logging.debug("Keeping directory tree at %s" % self.name)   
+            if settings.schroot:
+                logging.debug("Keeping schroot session %s at %s" % (self.schroot_session, self.name))
+            else:
+                logging.debug("Keeping directory tree at %s" % self.name)   
 
     def create_temp_tgz_file(self):
         """Return the path to a file to be used as a temporary tgz file"""
@@ -803,6 +816,13 @@ class Chroot:
             prefix.append('eatmydata')
         run(prefix + ["tar", "-C", self.name, "-zxf", tarball])
 
+    def setup_from_schroot(self, schroot):
+        self.schroot_session = schroot.split(":")[1] + "-" + str(uuid.uuid1()) + "-piuparts"
+        run(['schroot', '--begin-session', '--chroot', schroot , '--session-name', self.schroot_session])
+        ret_code, output = run(['schroot', '--chroot', self.schroot_session, '--location'])
+        self.name = output.strip()
+        logging.info("New schroot session in '%s'" % self.name);
+
     def setup_from_lvm(self, lvm_volume):
         """Create a chroot by creating an LVM snapshot."""
         self.lvm_base = os.path.dirname(lvm_volume)
@@ -820,7 +840,11 @@ class Chroot:
         if settings.eatmydata and os.path.isfile(os.path.join(self.name,
                                                  'usr/bin/eatmydata')):
             prefix.append('eatmydata')
-        return run(["chroot", self.name] + prefix + command,
+        if settings.schroot:
+            return run(["schroot", "--preserve-environment", "--run-session", "--chroot", self.schroot_session, "--directory", "/", "-u", "root", "--"] + prefix + command,
+                   ignore_errors=ignore_errors)
+        else:
+            return run(["chroot", self.name] + prefix + command,
                    ignore_errors=ignore_errors)
 
     def create_apt_sources(self, distro):
@@ -2282,6 +2306,10 @@ def parse_command_line():
                       default="1G", help="Use SNAPSHOT-SIZE as snapshot size when creating " +
                       "a new LVM snapshot (default: 1G)")
 
+    parser.add_option("--schroot", metavar="SCHROOT-NAME", action="store",
+                      help="Use schroot session named SCHROOT-NAME for the chroot, instead of building " +
+                           "a new one with debootstrap.")
+
     parser.add_option("-m", "--mirror", action="append", metavar="URL",
                       default=[],
                       help="Which Debian mirror to use.")
@@ -2415,6 +2443,7 @@ def parse_command_line():
     settings.savetgz = opts.save
     settings.lvm_volume = opts.lvm_volume
     settings.lvm_snapshot_size = opts.lvm_snapshot_size
+    settings.schroot = opts.schroot
     settings.end_meta = opts.end_meta
     settings.save_end_meta = opts.save_end_meta
     settings.skip_minimize = opts.skip_minimize
